refactor(profiler): allocate a new tracker outside the lock
We don't currently instrument the raw allocator domain, but if we did,
then we wouldn't want to allocate it with the memalloc lock held or we
would miss the allocation in our profiles. Plus it's preferable to avoid
any reentrancy or deadlock concerns if we can help it. Allocate the new
tracker outside the lock in iter_events, and just swap out the global
alloc tracker pointer under the lock.