feat(ci_visibility): make pytest v2 plugin the default version (#12066)
Previously, the pytest v2 plugin was disabled by default, and could be
enabled by setting the `DD_PYTEST_USE_NEW_PLUGIN_BETA` variable to true.
In ddtrace 3.0, the pytest v2 plugin will be *enabled* by default, and
`DD_PYTEST_USE_NEW_PLUGIN_BETA` will not be used anymore.

For now I'm not removing the code for the old plugin version. Instead,
I'm adding an internal environment variable
`_DD_PYTEST_USE_LEGACY_PLUGIN` that allows running pytest with the old
plugin, and using that in the tests that depend on it. We can remove the
code (and the variable) after the 3.x release, since this is not part of
the public interface.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)