fix(profiling): fix crash in heap profiler due to unsigned integer wraparound
The traceback array type used by the heap profiler is instantiated from
a preprocessor macro in ddtrace/profiling/collector/_utils.h. The macro
takes an index/count type, which for the traceback array was a uint16_t.
If the heap profiler's traceback array gets large, meaning we're
tracking a lot of allocations, then computing the new size of the array
when growing it can easily overflow the size type. This can lead us to
allocate a smaller array rather than a larger one, and yet the count of
elements stays the same. So we can write out of bounds, free NULL
pointers, etc, when we access elements outside the bounds of the actual
array.

Change the size type to a size_t for that array. The other arrays
instantiated from that macro already use 64-bit unsigned integers, which
are at least as big as size_t on our supported platforms. Really we
could do away with the size type parameter and just always use size_t.
But the next step for this code is probably to move it to C++ or Rust,
where we can use an existing dynamic array type, rather than to muck
with that set of macros.

This change also adds an assert when growing the array that the new size
doesn't wrap around. I originally found the bug by running this program:

	from ddtrace.profiling.collector import _memalloc

	_memalloc.start(128, 10000, 8)

	count = 100_000
	thing_size = 32

	junk = []
	for i in range(count):
		b1 = bytearray(thing_size)
		b2 = bytearray(2*thing_size)
		b3 = bytearray(3*thing_size)
		b4 = bytearray(4*thing_size)
		t = (b1, b2, b3, b4)
		junk.append(t)

	_memalloc.stop()

	del junk

Prior to this fix, it crashes when growing the array, as described.
After the fix, it runs to completion without crashing. However, it takes
~50 seconds to run on my macbook due to the O(N) lookup when untracking
an element, which leads to O(N^2) cost to untrack all the allocations at
the end of the program. So, reviewers can manually run this to confirm
the fix but I'm hesistant to add a 50 second program to our already slow
CI.