feat(tracing): add support for inferred proxy spans to support aws api gateway tracing (#12053)
## Motivation

This PR adds support for Inferred Proxy spans. Inferred proxies
supported with this PR are AWS API Gateway.

To support these inferred spans, a few steps were taken.
- All Web Server integrations were previously refactored to use the same
`ddtrace.core._start_span` trace handler in a previous PR.
- In this PR, before starting the `web_server.request` span, we look for
the existence of a few headers that the request may contain, which
signal that the request came from an inferred proxy system that we
support (in this case AWS API Gateway).
- If the request contains all necessary Inferred Proxy headers, an
"inferred span" is created with the information contained within the
headers. This span becomes our root span, and the web server request
span is the child of this inferred span.
- The PR takes into account that the inbound request may contain
distributed trace information, and maintains this functionality, the
only change being that the distributed child span is now the inferred
proxy span, and not the webserver request span, which is now the
grandchild span.
- If all necessary headers are not received (see code implementation for
these headers), then the web server span is created as normal.

Implementation Details:
- The feature is disabled by default.
- The inferred proxy span inherits any downstream errors that occur
within the web server request span.