ci(serverless): tests for all binaries removed in sls builds. (#12168)
There are several binary files that are removed from ddtrace when it is
packaged as a AWS Lambda Layer (see
https://github.com/DataDog/datadog-lambda-python/blob/main/Dockerfile).
This PR adds testing to ensure that none of those files are imported in
certain cases.

Note that the `ddtrace.contrib.requests` package currently attempts to
import a removed file. This test has been marked xfail.