refactor(profiling): add accuracy tests for heap profiling
There aren't any tests of the statistical accuracy of the heap profiler.
Start adding some. This commit adds two tests, though they don't
currently have any explicit assertions and aren't yet plugged in.
They're intended to be run while working on the heap profiler, to have a
way of examining the profiler's output and performance while it's being
actively worked on. And they'll eventually turn into tests we actually
run in CI.

One test allocates known amounts of memory in different functions, keeps
those allocations alive so they'll be captured in a heap profile, and
compares the values reported in the heap profile to the actual amount of
memory allocated according to tracemalloc. This is repeated for several
different sampling rates. The test doesn't yet make any assertions.
We'll probably want to think harder about the sampling method and
upscaling and what kind of variance we expect, in order to make
reasonable, non-flaky assertions.

The second test is a simplified version of the first, intended to be run
with high sampling, that will be really really slow with our current
implementation. The test program used to crash. It doesn't now, but it's
too slow to be worth running in CI. But one goal of my work is to make
sure the overhead of the profiler is reasonable, especially the scaling
behavior for tracking lots of allocations. If the performance improves,
we can run it as a regression test.