fix: prevent installation of incompatible envier==0.5.1 (#11286)
Use of EnvVariable._cast was introduced in commit 48c2a2c, released in 2.15.1, and the requirement for ddtrace
was set to ~=0.5 back in early 2024 in 00a7f00. This is generally fine but as seen in the issue it's sporadically
possible to end up with envier==0.5.1 in your environment and cause ddtrace to fail.