refactor(profiling): document heap profile sampling (#12483)
The heap profiler does statistical sampling of allocations. There is no
explanation in the code (or elsewhere) how the sampling in the profiler
works or why the chosen method is justified. We want to know whether
it's fair and whether our reported numbers accurately represent the real
heap size and the relative portion of the heap taken by different
objects of varying sizes.

Prior to working on this code, I was particularly concerned about the
weighting of sampled values. The method this profiler uses is different
than the methods used by either the Go profiler or tcmalloc, which seem
to generally do a good job. I've done some testing which seems to
indicate that the weighting we do is actually pretty good. So this
commit documents _why_ it's okay. My model of sampling borrows pretty
heavily from tcmalloc's documentation, which I think does the best job
of describing how it ought to work out of any resource I've found so
far. I also added a comment describing how the next sampling point is
chosen, since it might not be obvious right away from looking at the
code how `heap_tracker_next_sample_size`'s math relates to the described
sampling method.