chore: use guess-next-dev instead of release-branch-semver [3.1] (#12499)
This PR updates the `version_schema` in the `../pyproject.toml` file for
the 3.1 release branch from `release-branch-semver` to `guess-next-dev`.
This is to ensure that system tests work as intended with backports to
this release branch.

IMPORTANT: This PR needs to be merged before the first backport is
created for 3.1.Otherwise, system tests will not work as expected.