fix(profiling): make reentrancy guard actually thread-local (#12526)
The memory profiler has a guard to avoid re-entering its code if there
are allocations during sampling. This guard was meant to be
thread-local, but was not correctly declared as such. This doesn't
immediately cause problems because the profiler uses try-locks to
protect access to its data structures, and re-entering the code will
fail to acquire the locks. But this bug could be a source of deadlocks
or data corruption if the code changes substantially in the future. This
fix makes the guard thread-local as originally intended.  Also document
a little more the thread-local access model we specify here, since it
might not be obvious what it's for or why we need to explicitly declare
it.

Note: this might increase the overhead of the memory profiler. global-dynamic
TLS access (the kind we should be/might implicitly end up using here) can
involve a few function calls for every access to the variable, and we access
the reentrancy guard for every allocation, whether or not we sample the
allocation.