ci: move `promote-oci-to-prod` job to release stage (#12534)
We want to ensure that no artifacts are published anywhere until the
very end of the release pipeline. This was true for our PyPI pubilshing,
but not for OCI images. This change will make the OCI publish job
dependent on no jobs, but part of the release stage to ensure this.

This was tested on [this
commit](https://gitlab.ddbuild.io/DataDog/apm-reliability/dd-trace-py/-/pipelines/57265203)
(temporarily added this feature branch as a valid trigger for release
stage), and it successfully blocked the OCI job when the PyPI publish
job failed.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [ ] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)