ci: move aiobotocore suite to gitlab (#12553)
This change moves the `aiobotocore` test suite to gitlab as part of the
ongoing migration away from circleci.

This change replaces the `motoserver` image used by this suite with one
from the official registry, as opposed to the fork that had been used
previously. It also makes the `moto` service configurable like the other
services used in dd-trace-py's test suite, and updates some test code to
account for the fact that newer versions of `motoserver` allow all mock
AWS services to be accessed via the same port on the same container.

Changes to S3 bucket naming in tests are a workaround for the "already
exists" errors that this new version of motoserver raises on repeated
calls to `create_bucket` for the same bucket name.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)