refactor(di): make DI depend on Symbol DB
We add an explicit dependency on Symbol DB to DI. The current
implementation of Symbol DB is relevant for just DI, so we make sure
this is reflected in the implementation. The added benefit is that we
can make some of the imports lazy and save startup time when DI is not
enabled.