ci(benchmarks): refactor microbenchmark jobs (#12495)
This PR introduces a few major changes to the `microbenchmarks` jobs in
GitLab:

1. Uses built wheels for the candidate instead of re-building them for
every benchmark
1. The goal here was two fold. 1) save on benchmark execution time by
using a "cached" build of the project instead of building for every
benchmark job and 2) test the exact same wheels that we build and ship
to customers.
2. Adds a new step to detect the appropriate baseline and builds or
downloads the appropriate wheel to share amongst all the benchmark jobs
1. One major change here is that we will always have a baseline to test
against. Previously if no PR was found then no baseline was used, so we
could not compare. However, now that we rely on the
`download_ddtrace_artifacts` then we will 100% have a PR, and worse case
we will fall back to `main` to test against when we cannot determine the
actual base.
2. One `main`, release branches, and release tags we will find the most
recent non-rc tag in that branch and download the wheel from PyPI to use
as the baseline to test against. This ensures that merges to `main` and
release tags also get comparison runs. This is a pre-requisite for
enabling (pre-)release alerting and blocking based on performance
overhead added by a given new release.
3. We always run the benchmark from the candidate branch.
1. Previously we would checkout the code for the baseline branch and
then only if the scenario existed there run the version in the baseline
branch. With these changes we always execute the benchmark scenario from
the code that exists in the candidate branch. This way if new benchmarks
are being added we can still get a comparison in the same PR adding
them, and if we make changes to how the benchmark is structured we can
compare apples-to-apples with the baseline rather than testing two
different benchmark configurations/scripts.


## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [ ] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

---------

Co-authored-by: Rachel Yang <rachel.yang@datadoghq.com>
Co-authored-by: Gabriele N. Tornetta <P403n1x87@users.noreply.github.com>
Co-authored-by: Munir Abdinur <munir_abdinur@hotmail.com>
Co-authored-by: Munir Abdinur <munir.abdinur@datadoghq.com>
Co-authored-by: Joey Zhao <5253430+joeyzhao2018@users.noreply.github.com>