chore(errortracking): automatically report handled exceptions through span events (#12561)
### Overview

**The feature is intended only to be dogfooded for now and will not be
directly publicly available.**

This PR aims at adding support for automatic reporting of handled
exceptions for Python.
This opt-in feature will report handled exceptions through span events
for user code and third party packages depending on the configuration.
The sources and the test files lie in their respective **errortracking**
folder.

### Implementation details:
- In python3.10 and python3.11 we are using bytecode injection. We are
using sys.monitor as soon as it is possible as bytecode injection is
considered unsafe (could potentially alter the behavior of the user
program).
- In python3.12, we are using sys.monitoring. Notes that sys.monitoring
call the callback for every handled exceptions including the python
internal ones. Therefore it adds a significant overhead: 10-15% on an
endpoint with a try/except inserting something in a SQLite Database.

Some requirements:
- Python internals and ddtrace handled exceptions are **not** reported. 
- If the same error is handled multiple times, only one span event will
be created.
- If an error is handled then thrown, the error will appear on the span
and not as a span event.

 The feature is controlled by 4 env variables:
- _DD_ERROR_TRACKING_REPORT_HANDLED_ERRORS_ENABLED_ : _all_ | _user_ |
_third_party_. Report every handled exceptions or only user code or
third party packages handled exceptions.
- _DD_ERROR_TRACKING_REPORT_HANDLED_ERRORS_ENABLED_MODULES_ : specify
the modules from which handled exceptions will be reported.

In this PR, there are two additional env variables:
- _DD_ERROR_TRACKING_REPORT_HANDLED_ERRORS_AFTER_UNHANDLED_: report
handled exceptions only if an unhandled one occurred during the span.
This is a test env to see if it is useful and will likely be removed.
- _DD_ERROR_TRACKING_REPORT_HANDLED_ERRORS_LOGGER_: as the moment, the
span event errors don't appear automatically in Error Tracking so they
need to be logged as well. It will be removed eventually. You need to
specify the name of the logger you want to use.

### Tests
- test_handled_exceptions_by_bytecode is only for python3.10 and
python3.11. It checks that the different types of except construct are
well supported by the bytecode injection.
- test_handled_exceptions is for everything. It tests the feature. Notes
that the definition of user/third_party_package is not very clear in the
test. Therefore external user code as well as a fake numpy package are
simulated.
Due to limitation, the main code cannot be instrumented for python3.10
and 3.11, that is why the functions to instrument lie in a different
file.

The bottle, fastAPI and Django tests were run against the feature here:
https://github.com/DataDog/dd-trace-py/actions/runs/13587303860

The django tests timed out but no error occurred.

## Checklist
- [X] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)